ActiveUser & AuditedObject issues
- ActiveUser cannot be injected into EntityListener in GF 4.1. It has
already been reported to GF and we have logged it in #26.
- AuditedObject is not a mapped class and therefore all its child
classes such as EnterpriseObject and EnterpriseRelationship are not
having the changed date fields generated in the DB tables. We have
changed AuditedObject to be an interface and all subclasses will store
their own fields while the interface only provides the methods to be
used in the AuditedObjectListener.